Support slice indexing on HeapReferenceList
This is a convenience for sending a partial list at the start or end
of transmission. It's not optimal because it performs O(n)
reference-counting operations (which is what HeapReferenceList was
created to avoid), but that's no worse than Python list slicing. The
alternative would be to build something akin to numpy's indexing, which
just gives a view of the base object, but that would add a lot of
complexity.