Implement support for generic receive offload
GRO allows a socket to opt in to having the kernel group together
packets with the same header information (source, destination, size etc)
into one jumbo packet. The original packet size is provided as ancillary
data to allow userspace to reconstruct the original packets. This can
provide around a 20% speedup.