Refactor passthrough unit tests
The tests were originally structured using inheritance to work with the
limitations of nosetests. This is the age of pytest, so restructure
things so that the transport becomes a parametrised fixture, and the
tests are orthogonal to the transports.

Having the asynchronous versions of the tests in a separate file was
difficult to do with this approach, so they're now all combined in
one file.

The interface between the transport and the tests is also modified so
that prepare_receivers can now dynamically allocate ports and pass those
ports on to prepare_senders. These tests no longer use hard-coded ports,
making them same to run concurrently.