Reduce some code duplication in transport tests
Make async methods invoke the methods on the sync versions of the class
where applicable. This is a little dirty because what's being passed
isn't an instance of the right class, but it works because Python is
duck-typed. It'll probably need revisiting if I ever get serious about
type annotations.