Change how features are force-enabled
Previously -Dauto_features was used, but with the addition of SVE, no
platform can have all features enabled. Instead, pass flags to
force-enable features that we want to ensure are enabled on a
per-platform basis. These are abstracted into a helper script to avoid
repeating them for each usage. This script also takes care of -Dwerror
and --native.