Update a number of Boost.Test assertions
Update to use BOOST_TEST where possible instead of BOOST_CHECK_EQUAL or
BOOST_CHECK. BOOST_TEST is more "modern" but doesn't cope with
expressions that aren't be formatted with ostream, so it's not applied
everywhere.

The main motivation for this is that BOOST_CHECK_EQUAL leads to useless
clang error messages when something goes wrong (it never reports the
assertion that is the problem). This has been showing up with
-Wsign-compare, and in some places the RHS of comparisons has been made
explicitly unsigned to address these errors.