Hack: remove check for packet size consistency
I don't intend this ever to be merged into main. This is just a quick
hack to test with skarab dsim inputs.

The SKARAB dsim packets are slightly too big due to the word length in
the skarab 40G core. So they don't match the digitiser packets
precisely. SKARAB F-engines just ignore this but GPU F-engines don't
becasue of the changes made in this commit.

I just need to test the routing mechanim properly and then I will delete
this branch.