Shift a couple of stages from Jenkinsfile into Dockerfile
The Jenkins job on our new deployment is running Docker containers with
the option `-u 1000:1000` instead of `-u 0:0` which results in
permissions issues when doing administrative things such as using
`apt-get` to install stuff.

So this commit moves the install stages of the test run into a
Dockerfile, so that the image which is all nicely configured can then be
used to run the tests.

Relates to NGC-1447.