Complete rework of discard_oldest functionality
The approach was broken because it destroyed items, but ChunkRingStream
is intended to never destroy elements, merely recycle them. So the
ringbuffer interface needs to return removed items.

For now I've completely reverted the Python API. It will need to be
developed once something has been added at the chunked receiver level.