Add discarded_chunks statistic for pop_if_full feature
This required some major rework of the way ready callbacks are handled.
It turned out the previous approach didn't carry enough information for
this to be workable for chunk stream groups, because the batch_stats
pointer is only useful if you know the statistic to slot mapping, and
that is specific to the stream rather than the group (even if it will
typically be uniform across the streams).

This thus required introducing a new variant on the chunk_ready_function
type (group_chunk_ready_function) and an awkward backwards compatibility
mechanism.

The same thing probably ought to be done for chunk_allocation_function,
but I've only partly gotten around to that.